# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keyloop Unified Service Scheduler",
    "description": "Service-appointment scheduling for automotive dealerships. No authentication (ADR-0002); a candidate list is advisory only — every write is adjudicated by PostgreSQL (CLAUDE.md §2.1).",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/appointments/{id}/cancellation": {
      "post": {
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A confirmed or cancelled appointment.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "A confirmed or cancelled appointment.",
                  "type": "object",
                  "required": [
                    "id",
                    "dealershipId",
                    "customerId",
                    "vehicleId",
                    "serviceTypeId",
                    "technicianId",
                    "bayId",
                    "startsAt",
                    "endsAt",
                    "status"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "dealershipId": {
                      "type": "string"
                    },
                    "customerId": {
                      "type": "string"
                    },
                    "vehicleId": {
                      "type": "string"
                    },
                    "serviceTypeId": {
                      "type": "string"
                    },
                    "technicianId": {
                      "type": "string"
                    },
                    "bayId": {
                      "type": "string"
                    },
                    "startsAt": {
                      "type": "string"
                    },
                    "endsAt": {
                      "type": "string"
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "confirmed"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "cancelled"
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem detail.",
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "properties": {
                    "type": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "/problems/malformed-request"
                          ]
                        }
                      ]
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "resource": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "bay"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "technician"
                          ]
                        }
                      ]
                    },
                    "reference": {
                      "type": "string"
                    },
                    "opensAt": {
                      "type": "string"
                    },
                    "closesAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem detail.",
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "properties": {
                    "type": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "/problems/appointment-not-found"
                          ]
                        }
                      ]
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "resource": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "bay"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "technician"
                          ]
                        }
                      ]
                    },
                    "reference": {
                      "type": "string"
                    },
                    "opensAt": {
                      "type": "string"
                    },
                    "closesAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}